requirements.txt
Django==1.6.1
Pillow==2.3.0
South==0.8.4
boto==2.24.0
dj-database-url==0.2.2
dj-static==0.0.5
django-grappelli==2.5.1
django-storages==1.1.8
gunicorn==18.0
psycopg2==2.5.1
sorl-thumbnail==11.12
static==0.4
wsgiref==0.1.2
